refactor: move posting timeline
This patch simply moves the PostingTimelineView into its own file outside of
ContentView.
